Move utility functions to utils.{c,h}
